requirements.txt
python-dotenv==1.0.1
discord.py==2.4.0
google-generativeai==0.7.2
urllib3==2.2.2

